MovieComments/requirements.txt
pymysql
lxml
requests
pyexecjs
